Move tx error into TransactionStatusIcon
