+++ src/app/components/TransactionStatusIcon/index.tsx
@@ -6,6 +6,7 @@ import CancelIcon from '@mui/icons-material/Cancel'
 import { styled } from '@mui/material/styles'
 import { COLORS } from '../../../styles/theme/colors'
 import HelpIcon from '@mui/icons-material/Help'
+import { TxError } from '../../../oasis-indexer/api'
 
 type TxStatus = 'unknown' | 'success' | 'failure'
 
@@ -48,16 +49,30 @@ const StyledBox = styled(Box, {
   }
 })
 
+const ErrorBox = styled(Box)(() => ({
+  display: 'flex',
+  justifyContent: 'center',
+  alignItems: 'center',
+  minHeight: '28px',
+  fontSize: '12px',
+  backgroundColor: COLORS.grayLight,
+  color: COLORS.errorIndicatorBackground,
+  borderRadius: 10,
+  paddingLeft: 12,
+  paddingRight: 12,
+}))
+
 type TransactionStatusIconProps = {
   /**
    * Did the transaction succeed?
    * A missing value means unknown. (For encrypted transactions).
    */
-  success?: boolean
+  success: undefined | boolean
+  error: undefined | TxError
   withText?: boolean
 }
 
-export const TransactionStatusIcon: FC<TransactionStatusIconProps> = ({ success, withText }) => {
+export const TransactionStatusIcon: FC<TransactionStatusIconProps> = ({ success, error, withText }) => {
   const { t } = useTranslation()
   const status: TxStatus = success === undefined ? 'unknown' : success ? 'success' : 'failure'
   const statusLabel: Record<TxStatus, string> = {
@@ -67,14 +82,21 @@ export const TransactionStatusIcon: FC<TransactionStatusIconProps> = ({ success,
   }
 
   return (
-    <StyledBox success={success} withText={withText}>
-      {withText && (
-        <span>
-          {statusLabel[status]}
-          &nbsp;
-        </span>
+    <>
+      <StyledBox success={success} error={error} withText={withText}>
+        {withText && (
+          <span>
+            {statusLabel[status]}
+            &nbsp;
+          </span>
+        )}
+        {statusIcon[status]}
+      </StyledBox>
+      {withText && error && (
+        <ErrorBox>
+          {error.message} ({t('errors.code')} {error.code})
+        </ErrorBox>
       )}
-      {statusIcon[status]}
-    </StyledBox>
+    </>
   )
 }
